Enum playground::shift_register::ShiftRegisterType

pub enum ShiftRegisterType {
    LED,
    RELAIS,
}

Representiert die verschiedenen Shift Register Typen

Zur Zeit gibt es 2 verschiedene Shift Register Typen

LED - Led Shift Register sind drei 8bit, serial in, paralel out, Shift Register, daisy chained - Nur die ersten 24 Ausgänge sind verbunden

RELAIS - Die Relais Shift Register sind zwei 8bit, serial in, paralel out, Shift Register, daisy chained - Nur die ersten 9 Ausgänge sind verbunden

Variants

LED

RELAIS