Struct playground::shift_register::ShiftRegister

pub struct ShiftRegister {
    pub data: u64,
    // some fields omitted
}

Datenstruktur der Shift Register Hardware

Das data Feld ist ein Buffer der den aktuellen Zustand der Shift Register wiederspiegelt. Shift Register können nur geschrieben werden, desshalb benötigt man ein Speicherbereich um zum Beispiel den aktuellen Zustand einzelner Bits abfragen zu können.

Fields

data

Methods

impl ShiftRegister

fn new(register_type: ShiftRegisterType) -> Self

Erzeugt ein neuen Shift Register

Arguments

register_type - Art des Shift Registers

Examples

use playground::shift_register::{ShiftRegister,ShiftRegisterType};

let led = ShiftRegister::new(ShiftRegisterType::LED);
let relais = ShiftRegister::new(ShiftRegisterType::RELAIS);
assert_eq!(led.data, 0b0);
assert_eq!(relais.data, 0b0);

fn set(&mut self, num: u64)

Setzt das übergebene Bit im Shift Register data Buffer

Arguments

num - Nummer des zu setzenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. Das bedeutet set(1) setzt das erste Bit(0) im data Buffer.

Examples

use playground::shift_register::{ShiftRegister,ShiftRegisterType};

let mut led = ShiftRegister::new(ShiftRegisterType::LED);
assert_eq!(led.data, 0b0);
led.set(3);
assert_eq!(led.data, 0b100);

More info: http://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit-in-c-c

fn get(&self, num: u64) -> bool

Abfrage ob ein Bit gesetzt ist, true wenn ja, false wenn das bit nicht gesetzt ist

Arguments

num - Nummer des abzufragenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. Das bedeutet get(1) fragt das erste Bit(0) im data Buffer ab.

Examples

use playground::shift_register::{ShiftRegister,ShiftRegisterType};

let mut led = ShiftRegister::new(ShiftRegisterType::LED);
led.set(1);
led.set(3);
assert_eq!(led.get(1), true);
assert_eq!(led.get(2), false);
assert_eq!(led.get(3), true);

More info: http://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit-in-c-c

fn clear(&mut self, num: u64)

Löscht das übergebene Bit

Arguments

num - Nummer des zu löschenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. Das bedeutet clear(1) löscht das erste Bit(0) im data Buffer.

Examples

use playground::shift_register::{ShiftRegister,ShiftRegisterType};

let mut led = ShiftRegister::new(ShiftRegisterType::LED);
assert_eq!(led.data, 0b0);

led.set(3);
assert_eq!(led.get(3), true);

led.clear(3);
assert_eq!(led.get(3), false);

fn toggle(&mut self, num: u64)

Schaltet das übergebene Bit um, war es Null dann wird es Eins und umgekehrt

Arguments

num - Nummer des zu wechselnden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. Das bedeutet toggle(1) schaltet das erste Bit(0) im data Buffer um.

Examples

use playground::shift_register::{ShiftRegister,ShiftRegisterType};

let mut led = ShiftRegister::new(ShiftRegisterType::LED);
assert_eq!(led.data, 0b0);

led.toggle(3);
assert_eq!(led.get(3), true);
led.toggle(3);
assert_eq!(led.get(3), false);